WritingAssistantInterface/pyproject.toml
[project]
name = "writingassistantinterface"
version = "0.1.0"
description = ""
authors = [
    {name = "Philippe-Lauwers",email = "philippe.lauwers@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "flask (>=3.1.1,<4.0.0)",
    "flask-cors (>=6.0.1,<7.0.0)",
    "requests (>=2.32.4,<3.0.0)",
    "flask-sqlalchemy (>=3.1.1,<4.0.0)",
    "flask-migrate (>=4.1.0,<5.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "flask-login (>=0.6.3,<0.7.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
